% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{check_disk_space}
\alias{check_disk_space}
\title{Check available disk space}
\usage{
check_disk_space(dir_path, required_gb = 50)
}
\arguments{
\item{dir_path}{Path to check}

\item{required_gb}{Minimum GB required}
}
\description{
Check available disk space
}
