% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_monthly_links.R
\name{combine_monthly_links}
\alias{combine_monthly_links}
\title{Combine monthly linked grid files}
\usage{
combine_monthly_links(month_YYYYMMs, link.to = "grids", filename = NULL)
}
\arguments{
\item{month_YYYYMMs}{months and years to combine. Format created by `get_yearmon()`}

\item{link.to}{spatial scale for plotting. Must be 'grids'.}

\item{filename}{What should the resulting RData file be called?
Defaults to `paste0('hyads_unwgted_', link.to, '.RData')`}
}
\value{
Saves an .RData file to the rdata_dir defined by `create_dirs()`
with filename `filename`.
}
\description{
`combine_monthly_links()` combines linked grid files produced with
`link_all_units()` into lists of data.tables for easier manipulating.
}
\details{
\code{combine_monthly_links}
}
