% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Progress_Hysplit_Simulation.R
\name{.obs_one}
\alias{.obs_one}
\title{Lecture obs d'un fichier avec cache (invalidé si mtime change)}
\usage{
.obs_one(f, mtime)
}
\description{
Lecture obs d'un fichier avec cache (invalidé si mtime change)
}
