% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{download_file}
\alias{download_file}
\title{Download meteorological and PBL data with disk space checking}
\usage{
download_file(url, file, dir)
}
\description{
Download meteorological and PBL data with disk space checking
}
