% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_met_reanalysis_https}
\alias{get_met_reanalysis_https}
\title{Download meteorological files from NOAA with robust error handling}
\usage{
get_met_reanalysis_https(
  files = NULL,
  years = NULL,
  path_met_files,
  max_retries = 3,
  timeout = 900
)
}
\description{
Download meteorological files from NOAA with robust error handling
}
